replace old test
